refactored code with data class and reusable Composables
