+++ app/src/main/java/com/codetutor/countryinfoapp/data/CountryInfo.kt
@@ -1,0 +1,12 @@
+package com.codetutor.countryinfoapp.data
+
+data class CountryInfo(val flagId: Int,
+                       val commonName: String,
+                       val nationalCapital: String,
+                       val officialName: String,
+                       val region: String,
+                       val subRegion: String,
+                       val currencySymbol: String,
+                       val currencyName: String,
+                       val mobileCode: String,
+                       val tld: String)
